Add Post Controller Views
