+++ app/controllers/posts_controller.rb
@@ -1,46 +1,9 @@
 class PostsController < ApplicationController
-  before_action :set_post, only: [ :show, :edit, :update, :destroy ]
-
   def index
     @posts = Post.all
   end
 
   def show
-  end
-
-  def new
-    @post = Post.new
-  end
-
-  def create
-    @post = Post.new(post_params)
-    if @post.save
-      redirect_to @post, notice: "Post was successfully created."
-    else
-      render :new, status: :unprocessable_entity, notice: "Post was not created."
-    end
-  end
-
-  def edit
-  end
-
-  def update
-    if @post.update(post_params)
-      redirect_to @post, notice: "Post was successfully updated."
-    else
-      render :edit, status: :unprocessable_entity, notice: "Post was not updated."
-    end
-  end
-
-  def destroy
-    @post.destroy!
-    redirect_to posts_url, notice: "Post was successfully destroyed."
-  end
-  private
-  def post_params
-    params.require(:post).permit(:content)
-  end
-  def set_post
     @post = Post.find(params[:id])
   end
 end
