+++ app/controllers/users_controller.rb
@@ -1,0 +1,30 @@
+class UsersController < ApplicationController
+  before_action :set_user, only: [:show, :edit, :update]
+
+  def index
+    @users = User.all
+  end
+
+  def show
+  end
+
+  def edit
+  end
+
+  def update
+    if @user.update(user_params)
+      redirect_to @user, notice: "User was successfully updated."
+    else
+      render :edit, status: :unprocessable_entity, alert: "User was not updated."
+    end
+  end
+
+  private
+  def set_user
+    @user = User.find(params[:id])
+  end
+
+  def user_params
+    params.require(:user).permit(:name, :email)
+  end
+end
