+++ app/views/users/posts/new.html.haml
