+++ app/views/users/show.html.haml
@@ -1,0 +1,33 @@
+%h1
+  = I18n.t('view.user_show')
+%p
+  = I18n.t('attributes.user.email')
+  %div
+    = @user.email
+%p
+  = I18n.t('attributes.user.name')
+  %div
+    - if @user.name.blank?
+      = I18n.t('attributes.blank')
+    - else
+      = @user.name
+%p
+  = I18n.t('attributes.user.profile')
+  %div
+    - if @user.profile.blank?
+      = I18n.t('attributes.blank')
+    - else
+      = @user.profile
+%p
+  = I18n.t('attributes.user.blog_url')
+  %div
+    - if @user.blog_url.blank?
+      = I18n.t('attributes.blank')
+    - else
+      = @user.blog_url
+%p
+  = I18n.t('attributes.user.created_at')
+  %div
+    = @user.created_at.strftime('%Y-%m-%d %H:%M:%S')
+= link_to 'Edit', edit_user_path(@user), class: 'btn btn-primary'
+= link_to 'Back', users_path, class: 'btn btn-secondary'
